Added the auto tuning

+++ src/main/java/org/pid4j/pid/DefaultPid.java
@@ -2,7 +2,7 @@ package org.pid4j.pid;
 
 public class DefaultPid extends AbstractPid implements Pid {
 //	private long lastTime = System.currentTimeMillis();
-	private Double sampleTime = 100.0;  // default = 0.1 seconds = 100 milliseconds
+	private int sampleTime = 100;  // default = 0.1 seconds = 100 milliseconds
 	
 	public Double compute(Double value) {
 		Double error = setPoint - value;
@@ -33,11 +33,11 @@ public class DefaultPid extends AbstractPid implements Pid {
 		return output;
 	}
 
-	public Double getSampleTime() {
+	public int getSampleTime() {
 		return sampleTime;
 	}
 
-	public void setSampleTime(Double sampleTime) {
+	public void setSampleTime(int sampleTime) {
 		this.sampleTime = sampleTime;
 	}
 }
